Add a path_prefix option
This is an idea that I would find useful in my application. I'm open to
other ideas on the exact implementation, but I basically want to allow
only certain parts of my application to use django-lti-authentication.
The default setup enables django-lti-authentication for every request,
but there are many cases where we only want to enable LTI auth for
certain routes, such as the LTI launch route.